+++ src/biotsavart_parallel.rs
@@ -2,7 +2,8 @@ use rayon::prelude::*;
 use std::num::NonZeroUsize;
 use std::thread::available_parallelism;
 
-use crate::biotsavart::{bfield_direct, hfield_direct_tet};
+use crate::biotsavart::{bfield_direct, hfield_direct_tet, hmag_direct_tet};
+use crate::vec3::Vec3;
 
 pub fn get_nthreads(nthreads_requested: u32) -> usize {
     let nthreads_available: usize = available_parallelism().unwrap_or(NonZeroUsize::MIN).get();
@@ -89,13 +90,52 @@ pub fn hfield_direct_tet_parallel(
     let _z = z.par_chunks(chunk_size);
     let _hx = hx.par_chunks_mut(chunk_size);
     let _hy = hy.par_chunks_mut(chunk_size);
-    let _bz = hz.par_chunks_mut(chunk_size);
+    let _hz = hz.par_chunks_mut(chunk_size);
 
-    (_x, _y, _z, _hx, _hy, _bz)
+    (_x, _y, _z, _hx, _hy, _hz)
         .into_par_iter()
         .try_for_each(|(_x, _y, _z, _hx, _hy, _hz)| {
             hfield_direct_tet(nodes_flat, vol, jdensity_flat, _x, _y, _z, _hx, _hy, _hz)
         })?;
 
+    Ok(())
+}
+
+pub fn hmag_direct_tet_parallel(
+    source_nodes: (&[f64], &[f64], &[f64]),
+    source_element_connectivity: &[[u32; 4]],
+    source_mvectors: &[Vec3],
+    target_nodes: (&[f64], &[f64], &[f64]),
+    target_element_connectivity: &[[u32; 4]],
+    hx: &mut [f64],
+    hy: &mut [f64],
+    hz: &mut [f64],
+    nthreads_requested: u32,
+) -> Result<(), ()> {
+    // TODO: length checks
+    let n: usize = target_nodes.0.len();
+    let nthreads: usize = get_nthreads(nthreads_requested);
+    let chunk_size: usize = (n / nthreads).max(1);
+
+    // chunk the inputs over target elements
+    let _tec = target_element_connectivity.par_chunks(chunk_size);
+    let _hx = hx.par_chunks_mut(chunk_size);
+    let _hy = hy.par_chunks_mut(chunk_size);
+    let _hz = hz.par_chunks_mut(chunk_size);
+
+    (_tec, _hx, _hy, _hz)
+        .into_par_iter()
+        .try_for_each(|(_tec, _hx, _hy, _hz)| {
+            hmag_direct_tet(
+                source_nodes,
+                source_element_connectivity,
+                source_mvectors,
+                target_nodes,
+                &_tec,
+                _hx,
+                _hy,
+                _hz,
+            )
+        })?;
     Ok(())
 }