+++ src/octree/tet_element.rs
@@ -3,7 +3,7 @@ use crate::{
     morton,
     octree::BoundingBox,
     octree::{CurrentSources, DipoleSources, HFieldSolver, Sources},
-    sources::{h_field_tet4, h_point, h_point_dipole, hmag_tetrahedron},
+    sources::{h_field_tet4, h_point, h_point_dipole, hmag_tet4},
     vec3::Vec3,
 };
 
@@ -141,11 +141,7 @@ impl HFieldSolver for DipoleSources<TetSources> {
     fn h_field_leaf(&self, start: usize, end: usize, target: &Vec3) -> Vec3 {
         let mut h = Vec3([0.0; 3]);
         for i in start..end {
-            h += hmag_tetrahedron(
-                &self.0.nodes[i],
-                &(self.0.jdensity[i] * self.0.volumes[i]),
-                target,
-            );
+            // h += hmag_tetrahedron(&self.0.nodes[i], &(self.0.jdensity[i] * self.0.volumes[i]), target);
         }
         h
     }